refactor(PlaceMap): Intent Handler 구조 표준화 및 패키지 정리
`PlaceMap` 화면의 사용자 액션(Action)과 이벤트(Event)를 처리하는 핸들러 클래스들을 `intent.handler` 패키지로 통합하고, 공통 인터페이스를 도입하여 구조를 표준화했습니다.

- **인터페이스 정의:**
    - `ActionHandler<ACTION, STATE>` 및 `EventHandler<EVENT>` 인터페이스를 새로 정의하여 핸들러의 구현 규약을 통일했습니다.

- **패키지 이동 및 구현 수정:**
    - 기존 `intent.action`과 `intent.event` 패키지에 있던 `SelectActionHandler`, `MapEventActionHandler`, `FilterActionHandler`, `PlaceMapEventHandler`, `MapControlEventHandler`를 `intent.handler` 패키지로 이동했습니다.
    - 각 핸들러 클래스가 `ActionHandler` 또는 `EventHandler` 인터페이스를 구현(`override`)하도록 로직을 수정했습니다.

- **의존성 주입(DI) 설정:**
    - `SelectActionHandler`, `MapEventActionHandler`, `FilterActionHandler`에 `@ContributesBinding(PlaceMapViewModelScope::class)`을 추가하여 의존성 주입 설정을 보강했습니다.
