+++ app/src/main/java/com/daedan/festabook/presentation/placeMap/intent/handler/ActionHandler.kt
@@ -1,0 +1,10 @@
+package com.daedan.festabook.presentation.placeMap.intent.handler
+
+import kotlinx.coroutines.flow.StateFlow
+
+interface ActionHandler<ACTION, STATE> {
+    val uiState: StateFlow<STATE>
+    val onUpdateState: ((before: STATE) -> STATE) -> Unit
+
+    suspend operator fun invoke(action: ACTION)
+}
