+++ app/src/main/java/com/daedan/festabook/presentation/placeMap/intent/handler/EventHandler.kt
@@ -1,0 +1,5 @@
+package com.daedan.festabook.presentation.placeMap.intent.handler
+
+interface EventHandler<EVENT> {
+    suspend operator fun invoke(event: EVENT)
+}
