+++ app/src/main/java/com/daedan/festabook/presentation/placeMap/intent/handler/MapEventActionHandler.kt
@@ -1,6 +1,8 @@
-package com.daedan.festabook.presentation.placeMap.intent.action
+package com.daedan.festabook.presentation.placeMap.intent.handler
 
+import com.daedan.festabook.di.placeMapHandler.PlaceMapViewModelScope
 import com.daedan.festabook.logging.DefaultFirebaseLogger
+import com.daedan.festabook.presentation.placeMap.intent.action.MapEventAction
 import com.daedan.festabook.presentation.placeMap.intent.event.MapControlEvent
 import com.daedan.festabook.presentation.placeMap.intent.event.PlaceMapEvent
 import com.daedan.festabook.presentation.placeMap.intent.state.LoadState
@@ -8,18 +10,21 @@ import com.daedan.festabook.presentation.placeMap.intent.state.PlaceMapUiState
 import com.daedan.festabook.presentation.placeMap.intent.state.await
 import com.daedan.festabook.presentation.placeMap.logging.PlaceBackToSchoolClick
 import com.daedan.festabook.presentation.placeMap.model.InitialMapSettingUiModel
+import dev.zacsweers.metro.ContributesBinding
 import dev.zacsweers.metro.Inject
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.StateFlow
 
 @Inject
+@ContributesBinding(PlaceMapViewModelScope::class)
 class MapEventActionHandler(
+    override val uiState: StateFlow<PlaceMapUiState>,
+    override val onUpdateState: ((before: PlaceMapUiState) -> PlaceMapUiState) -> Unit,
     private val _mapControlUiEvent: Channel<MapControlEvent>,
     private val _placeMapUiEvent: Channel<PlaceMapEvent>,
-    private val uiState: StateFlow<PlaceMapUiState>,
     private val logger: DefaultFirebaseLogger,
-) {
-    suspend operator fun invoke(action: MapEventAction) {
+) : ActionHandler<MapEventAction, PlaceMapUiState> {
+    override suspend operator fun invoke(action: MapEventAction) {
         when (action) {
             is MapEventAction.OnMapReady -> {
                 _mapControlUiEvent.send(MapControlEvent.InitMap)
